+++ example/windows/flutter/generated_plugins.cmake
@@ -4,6 +4,8 @@
 
 list(APPEND FLUTTER_PLUGIN_LIST
   flutter_js
+  share_plus
+  smart_auth
   url_launcher_windows
 )