+++ lib/src/Utils/dui_widget_registry.dart
@@ -27,12 +27,17 @@ import '../core/builders/dui_icon_button_builder.dart';
 import '../core/builders/dui_image_builder.dart';
 import '../core/builders/dui_listview_builder.dart';
 import '../core/builders/dui_lottie_builder.dart';
+import '../core/builders/dui_paginated_listview_builder.dart';
 import '../core/builders/dui_pin_field_builder.dart';
+import '../core/builders/dui_refresh_indicator_builder.dart';
 import '../core/builders/dui_rich_text_builder.dart';
+import '../core/builders/dui_safearea_builder.dart';
 import '../core/builders/dui_scaffold_builder.dart';
 import '../core/builders/dui_sized_box_builder.dart';
 import '../core/builders/dui_spacer_builder.dart';
 import '../core/builders/dui_stack_builder.dart';
+import '../core/builders/dui_stepper_builder.dart';
+import '../core/builders/dui_stream_builder.dart';
 import '../core/builders/dui_switch_builder.dart';
 import '../core/builders/dui_tab_view_builder.dart';
 import '../core/builders/dui_tab_view_item_builder.dart';
@@ -116,6 +121,7 @@ class DUIWidgetRegistry {
         DUITextFormFieldBuilder.create(data),
     'digia/stepper': (data, {registry}) =>
         DezervStepperBuilder.create(data, registry: DUIWidgetRegistry.shared),
+    'digia/flutterStepper': withoutRegistry(DUIStepperBuilder.create),
     'digia/dezerv/dialPad': (data, {registry}) =>
         DUIDezervDialPadBuilder.create(data),
     'digia/futureBuilder': (data, {registry}) => DUIFutureBuilder(data: data),
@@ -127,10 +133,16 @@ class DUIWidgetRegistry {
     'digia/pinField': withoutRegistry(DUIPinFieldBuilder.create),
     'digia/calendar': withoutRegistry(DUICalendarBuilder.create),
     'digia/lineChart': withoutRegistry(DUIChartBuilder.create),
+    'digia/paginatedListView': (data, {registry}) =>
+        DUIPaginatedListViewBuilder.create(data),
     'digia/timer': withoutRegistry(DUITimerBuilder.create),
     'digia/conditionalBuilder': (data, {registry}) =>
         DUIConditionalBuilderBuilder.create(data,
-            registry: DUIWidgetRegistry.shared)
+            registry: DUIWidgetRegistry.shared),
+    'digia/streamBuilder': (data, {registry}) => DUIStreamBuilder(data: data),
+    'digia/refreshIndicator': (data, {registry}) =>
+        DUIRefreshIndicator(data: data),
+    'digia/safeArea': (data, {registry}) => DUISafeAreaBuilder(data: data)
   };
 
   static const DUIWidgetRegistry shared = DUIWidgetRegistry();