+++ lib/src/components/dui_tab_view/dui_tab_view_props.g.dart
@@ -9,6 +9,11 @@ part of 'dui_tab_view_props.dart';
 DUITabViewProps _$DUITabViewPropsFromJson(Map<String, dynamic> json) =>
     DUITabViewProps(
       hasTabs: json['hasTabs'] as bool?,
+      indicatorSize: json['indicatorSize'] as String?,
+      initialIndex: json['initialIndex'],
+      labelPadding: json['labelPadding'] == null
+          ? null
+          : DUIInsets.fromJson(json['labelPadding']),
       selectedLabelColor: json['selectedLabelColor'] as String?,
       selectedLabelStyle: DUITextStyle.fromJson(json['selectedLabelStyle']),
       unselectedLabelColor: json['unselectedLabelColor'] as String?,
@@ -34,6 +39,9 @@ Map<String, dynamic> _$DUITabViewPropsToJson(DUITabViewProps instance) =>
       'dividerHeight': instance.dividerHeight,
       'isScrollable': instance.isScrollable,
       'indicatorColor': instance.indicatorColor,
+      'indicatorSize': instance.indicatorSize,
       'viewportFraction': instance.viewportFraction,
       'tabBarPadding': instance.tabBarPadding,
+      'labelPadding': instance.labelPadding,
+      'initialIndex': instance.initialIndex,
     };