+++ lib/src/components/dui_widget_creator_fn.dart
@@ -1,8 +1,9 @@
+import 'dart:ui';
+
 import 'package:flutter/material.dart';
 
 import '../Utils/basic_shared_utils/color_decoder.dart';
 import '../Utils/basic_shared_utils/dui_decoder.dart';
-import '../Utils/basic_shared_utils/lodash.dart';
 import '../Utils/basic_shared_utils/num_decoder.dart';
 import '../Utils/extensions.dart';
 import '../Utils/util_functions.dart';
@@ -10,6 +11,7 @@ import '../core/action/action_handler.dart';
 import '../core/action/action_prop.dart';
 import '../core/evaluator.dart';
 import 'utils/DUIStyleClass/dui_style_class.dart';
+import 'utils/dottedBorderWrapper.dart';
 
 // ignore: non_constant_identifier_names
 Widget wrapInContainer(
@@ -26,6 +28,11 @@ Widget wrapInContainer(
       DUIDecoder.toBorderRadius(styleClass.border?.borderRadius?.toJson());
   final height = styleClass.height?.toHeight(context);
   final width = styleClass.width?.toWidth(context);
+  final borderWidth = styleClass.border?.borderWidth;
+  final borderColor = styleClass.border?.borderColor ?? '#000000';
+  final borderType = styleClass.border?.borderStyle;
+
+  final alignment = DUIDecoder.toAlignment(styleClass.alignment);
   // final clipBehavior = DUIDecoder.toClip(styleClass.clipBehavior);
 
   // Probably unnecessary Optimisation:
@@ -36,21 +43,32 @@ Widget wrapInContainer(
       border == null &&
       borderRadius.isZero() &&
       height == null &&
-      width == null) {
+      width == null &&
+      alignment == null) {
     return child;
   }
 
-  return Container(
-    width: width,
-    height: height,
-    padding: padding,
-    margin: margin,
-    decoration: BoxDecoration(
-        color: bgColor.letIfTrue(toColor),
-        border: border,
-        borderRadius: borderRadius),
-    clipBehavior: !borderRadius.isZero() ? Clip.hardEdge : Clip.none,
-    child: child,
+  return Padding(
+    padding: margin,
+    child: DottedBorderWrapper(
+      borderRadius: borderRadius,
+      borderWidth: borderWidth,
+      color: makeColor(borderColor),
+      borderType: borderType ?? '',
+      child: Container(
+        width: width,
+        height: height,
+        padding: padding,
+        alignment: alignment,
+        decoration: BoxDecoration(
+          color: makeColor(bgColor),
+          // border: border,
+          borderRadius: borderRadius,
+        ),
+        clipBehavior: !borderRadius.isZero() ? Clip.hardEdge : Clip.none,
+        child: child,
+      ),
+    ),
   );
 }