+++ lib/src/core/functions/web_js_functions.dart
@@ -15,11 +15,17 @@ class WebJsFunctions implements JSFunctions {
   }
 
   @override
-  fetchJsFile(String path) {
-    ScriptElement script = ScriptElement()
-      ..src = path
-      ..type = 'text/javascript';
-    document.head?.append(script);
+  Future<bool> initFunctions(FunctionInitStrategy strategy) {
+    switch (strategy) {
+      case PreferRemote(remotePath: String remotePath):
+        ScriptElement script = ScriptElement()
+          ..src = '$remotePath?t=${DateTime.now().millisecondsSinceEpoch}'
+          ..type = 'text/javascript';
+        document.head?.append(script);
+        return Future.value(true);
+      case PreferLocal():
+        throw Exception('Local strategy not available for web');
+    }
   }
 }